task: moved instance avatar to its own component
